# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.CreatePageReq:
    properties:
      change_note:
        maxLength: 500
        type: string
      content:
        additionalProperties: {}
        type: object
      content_text:
        type: string
      parent_id:
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      title:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  entity.Page:
    properties:
      author:
        $ref: '#/definitions/entity.UserShort'
      author_id:
        type: string
      children:
        items:
          $ref: '#/definitions/entity.Page'
        type: array
      content:
        additionalProperties: {}
        type: object
      content_text:
        type: string
      created_at:
        type: string
      current_version:
        type: integer
      id:
        type: string
      last_editor:
        $ref: '#/definitions/entity.UserShort'
      last_editor_id:
        type: string
      parent_id:
        type: string
      position:
        type: integer
      space_id:
        type: string
      status:
        description: draft | published
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  entity.PageTree:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.PageTree'
        type: array
      id:
        type: string
      parent_id:
        type: string
      position:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  entity.PageVersion:
    properties:
      author:
        $ref: '#/definitions/entity.UserShort'
      author_id:
        type: string
      change_note:
        type: string
      content:
        additionalProperties: {}
        type: object
      content_text:
        type: string
      created_at:
        type: string
      id:
        type: string
      page_id:
        type: string
      title:
        type: string
      version:
        type: integer
    type: object
...
  entity.UpdatePageReq:
    properties:
      change_note:
        maxLength: 500
        type: string
      content:
        additionalProperties: {}
        type: object
      content_text:
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      title:
        maxLength: 500
        minLength: 1
        type: string
    type: object
...
  entity.UserShort:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
...
  v1.movePageReq:
    properties:
      parent_id:
        type: string
      position:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Project management backend API (Jira-like).
  title: JiraFlow API
  version: "1.0"
paths:
...
  /api/v1/page-versions/{id}:
    get:
      parameters:
      - description: Page version ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.PageVersion'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get page version by ID
      tags:
      - page-versions
  /api/v1/pages:
    get:
      parameters:
      - description: Filter by space
        in: query
        name: space_id
        type: string
      - description: Filter by parent page
        in: query
        name: parent_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.Page'
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                type: integer
              total_pages:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: List pages
      tags:
      - pages
  /api/v1/pages/{id}:
    delete:
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete page
      tags:
      - pages
    get:
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Page'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get page by ID
      tags:
      - pages
    put:
      consumes:
      - application/json
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Page'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update page
      tags:
      - pages
  /api/v1/pages/{id}/move:
    put:
      consumes:
      - application/json
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      - description: New position data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.movePageReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move page (change position or parent)
      tags:
      - pages
  /api/v1/pages/{page_id}/versions:
    get:
      parameters:
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.PageVersion'
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                type: integer
              total_pages:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: List all versions of a page
      tags:
      - page-versions
  /api/v1/pages/{page_id}/versions/{version}:
    get:
      parameters:
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.PageVersion'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get specific version of a page
      tags:
      - page-versions
...
  /api/v1/spaces/{space_id}/pages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Space ID
        in: path
        name: space_id
        required: true
        type: string
      - description: Page data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreatePageReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Page'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create page in space
      tags:
      - pages
  /api/v1/spaces/{space_id}/pages/tree:
    get:
      parameters:
      - description: Space ID
        in: path
        name: space_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.PageTree'
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get page tree for space
      tags:
      - pages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"